requirements.txt
accelerate
datasets
fire
transformers>=4.51.3
torch
jsonlines
libclang
tree_sitter
tree_sitter_cpp
requests==2.27.1
lark
# flash-attn==1.0.4 # Depends on the GPU and CUDA version
